requirements.txt
python-telegram-bot
flask
python-dotenv
pyngrok
requests
gunicorn
